scroll-lock(nit): consolidate utils
